+++ src/public/assets/scripts/index.js
@@ -1,0 +1,7 @@
+// @prepros-append ./initializer.js
+// @prepros-append ./utils.js
+// @prepros-append ./split-screen.js
+// @prepros-append ./info-box.js
+// @prepros-append ./modal-handler.js
+// @prepros-append ./main.js
+// @prepros-append ./listeners.js